requirements.txt
openai
google-genai
python-dotenv
mem0ai
sentence-transformers
pytest
